+++ src/test/groovy/nebula/plugin/dependencylock/DependencyLockLauncherSpec.groovy
@@ -207,7 +207,6 @@ class DependencyLockLauncherSpec extends IntegrationSpec {
         def result = runTasksSuccessfully('dependencies')
 
         then:
-        println result.standardOutput
         result.standardOutput.contains 'test.example:foo:1.0.1 -> 1.0.0'
     }
 
@@ -222,7 +221,6 @@ class DependencyLockLauncherSpec extends IntegrationSpec {
         def result = runTasksSuccessfully('gL', 'dependencies')
 
         then:
-        println result.standardOutput
         !result.standardOutput.contains('test.example:foo:1.0.1 -> 1.0.0')
     }
 
@@ -237,8 +235,76 @@ class DependencyLockLauncherSpec extends IntegrationSpec {
         def result = runTasksSuccessfully(':generateLock', 'dependencies')
 
         then:
-        println result.standardOutput
         !result.standardOutput.contains('test.example:foo:1.0.1 -> 1.0.0')
+    }
+
+    @Issue('#79')
+    def 'lock file is applied on subprojects not being locked while generating lock with qualified task name'() {
+        setupCommonMultiproject()
+        new File(projectDir, 'sub1/dependencies.lock').text = PRE_DIFF_FOO_LOCK
+        new File(projectDir, 'sub2/dependencies.lock').text = PRE_DIFF_FOO_LOCK
+
+        when:
+        def result = runTasksSuccessfully('sub1:generateLock', ':sub2:dependencies')
+
+        then:
+        result.standardOutput.contains('test.example:foo:1.+ -> 1.0.0')
+    }
+
+    @Issue('#79')
+    def 'lock file ignored in multiproject on specific project with no leading :'() {
+        setupCommonMultiproject()
+        new File(projectDir, 'sub1/dependencies.lock').text = PRE_DIFF_FOO_LOCK
+        new File(projectDir, 'sub2/dependencies.lock').text = PRE_DIFF_FOO_LOCK
+
+        when:
+        def result = runTasksSuccessfully('sub1:generateLock', 'sub1:dependencies')
+
+        then:
+        !result.standardOutput.contains('test.example:foo:2.0.0 -> 1.0.0')
+    }
+
+    @Issue('#79')
+    def 'lock file ignored in nested multiproject when generating locks'() {
+        def middleDir = addSubproject('middle')
+        def sub0Dir = new File(middleDir, 'sub0')
+        sub0Dir.mkdirs()
+        new File(sub0Dir, 'build.gradle').text = """\
+            apply plugin: 'java'
+            dependencies {
+                compile 'test.example:foo:2.0.0'
+            }
+        """.stripIndent()
+        new File(sub0Dir, 'dependencies.lock').text = PRE_DIFF_FOO_LOCK
+        def sub1Dir = new File(middleDir, 'sub1')
+        sub1Dir.mkdirs()
+        new File(sub1Dir, 'build.gradle').text = """\
+            apply plugin: 'java'
+            dependencies {
+                compile 'test.example:foo:1.+'
+            }
+        """.stripIndent()
+
+        buildFile << """\
+            allprojects {
+                ${applyPlugin(DependencyLockPlugin)}
+                group = 'test'
+            }
+            subprojects {
+                repositories { maven { url '${Fixture.repo}' } }
+            }
+        """.stripIndent()
+
+        settingsFile << '''\
+            include ':middle:sub0'
+            include ':middle:sub1'
+        '''.stripIndent()
+
+        when:
+        def result = runTasksSuccessfully(':middle:sub0:generateLock', ':middle:sub0:dependencies')
+
+        then:
+        !result.standardOutput.contains('test.example:foo:2.0.0 -> 1.0.0')
     }
 
     def 'override lock file is applied'() {
@@ -452,7 +518,6 @@ class DependencyLockLauncherSpec extends IntegrationSpec {
         def s = runTasksSuccessfully('-PdependencyLock.overrideFile=test.lock', 'generateLock', 'saveLock')
 
         then:
-        println s.standardOutput
         new File(projectDir, 'dependencies.lock').text == NEW_FOO_LOCK
     }
 
@@ -548,7 +613,6 @@ class DependencyLockLauncherSpec extends IntegrationSpec {
         def aa = runTasksSuccessfully('-PdependencyLock.overrideFile=override.lock', 'generateLock', 'saveLock')
 
         then:
-        println aa.standardOutput
         String lockText1 = '''\
             {
                 "compile": {
@@ -691,7 +755,6 @@ class DependencyLockLauncherSpec extends IntegrationSpec {
         def s = runTasksSuccessfully('generateGlobalLock')
 
         then:
-        println s.standardOutput
         String globalLockText = '''\
             {
                 "_global_": {
@@ -852,7 +915,6 @@ class DependencyLockLauncherSpec extends IntegrationSpec {
         def s = runTasksSuccessfully('generateGlobalLock')
 
         then:
-        println s.standardOutput
         String globalLockText = '''\
             {
                 "_global_": {
@@ -903,8 +965,6 @@ class DependencyLockLauncherSpec extends IntegrationSpec {
                     }
                 }
             }'''.stripIndent()
-
-        println s.standardOutput
 
         new File(projectDir, 'global.lock').text == globalLockText
     }
@@ -989,6 +1049,52 @@ class DependencyLockLauncherSpec extends IntegrationSpec {
                 }
             }'''.stripIndent()
         new File(projectDir, 'sub2/dependencies.lock').text == lockText2
+    }
+
+    def 'generateGlobalLock ignores existing global lock file'() {
+        setupCommonMultiproject()
+        new File(projectDir, 'global.lock').text = '''\
+            {
+                "_global_": {
+                    "test.example:foo": {
+                        "locked": "1.0.1",
+                        "transitive": [
+                            "test:sub1",
+                            "test:sub2"
+                        ]
+                    },
+                    "test:sub1": {
+                        "project": true
+                    },
+                    "test:sub2": {
+                        "project": true
+                    }
+                }
+            }'''.stripIndent()
+        String globalLockText = '''\
+            {
+                "_global_": {
+                    "test.example:foo": {
+                        "locked": "2.0.0",
+                        "transitive": [
+                            "test:sub1",
+                            "test:sub2"
+                        ]
+                    },
+                    "test:sub1": {
+                        "project": true
+                    },
+                    "test:sub2": {
+                        "project": true
+                    }
+                }
+            }'''.stripIndent()
+
+        when:
+        def r = runTasksSuccessfully('generateGlobalLock')
+
+        then:
+        new File(projectDir, 'build/global.lock').text == globalLockText
     }
 
     def 'throw exception when saving global lock, if individual locks are present'() {